# apps/web/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ASKO API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/notifications": {
      "get": {
        "operationId": "NotificationController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "unreadOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "group",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "readStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedNotificationsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationController_markAsRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "post": {
        "operationId": "NotificationController_markAllAsRead",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/{id}": {
      "delete": {
        "operationId": "NotificationController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/preferences": {
      "get": {
        "operationId": "NotificationController_getPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "put": {
        "operationId": "NotificationController_updatePreferences",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationPreferencesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreferencesResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/notifications/push-subscriptions": {
      "post": {
        "operationId": "NotificationController_registerPushSubscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterPushSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionResponseDto"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationController_unregisterPushSubscription",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "get": {
        "operationId": "NotificationController_listPushSubscriptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscriptionListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/address": {
      "post": {
        "operationId": "AddressController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      },
      "get": {
        "operationId": "AddressController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
    "/address/{id}": {
      "put": {
        "operationId": "AddressController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      },
      "delete": {
        "operationId": "AddressController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Addresses"
        ]
      },
      "get": {
        "operationId": "AddressController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressRecordDto"
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
    "/address/{id}/primary": {
      "patch": {
        "operationId": "AddressController_setPrimary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
    "/address/user/{userId}": {
      "get": {
        "operationId": "AddressController_findByUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressListResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Addresses"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyResponseDto": {
        "type": "object",
        "properties": {}
      },
...
      "NotificationType": {
        "type": "string",
        "enum": [
          "repair_status_changed",
          "repair_assigned",
          "repair_transferred_to_repairer",
          "repair_transferred_from_repairer",
          "repair_transferred_client",
          "repair_diagnostics_declined",
          "repair_completed",
          "invoice_created",
          "payment_paid",
          "payment_failed",
          "payment_refunded",
          "invoice_unpaid_reminder",
          "repair_assignment_reminder",
          "repair_in_progress_stuck",
          "certificate_issued",
          "certificate_expiring_soon",
          "certificate_expired",
          "chat_message",
          "chat_conversation_created",
          "chat_participant_added",
          "chat_participant_removed",
          "message",
          "system",
          "schedule_created",
          "schedule_approved",
          "schedule_rejected",
          "schedule_updated",
          "schedule_deleted",
          "schedule_extra_day_requested",
          "schedule_extra_day_accepted",
          "schedule_extra_day_rejected",
          "schedule_pattern_created",
          "schedule_pattern_updated",
          "schedule_pattern_deleted",
          "schedule_pattern_approved",
          "schedule_pattern_rejected",
          "avr_signing_requested",
          "avr_signed",
          "repair_schedule_ending",
          "repair_schedule_auto_paused",
          "address_validated",
          "address_validation_failed",
          "user_device_validated",
          "user_device_validation_failed",
          "certificate_integrity_failed",
          "repair_part_shipped"
        ]
      },
      "NotificationTargetType": {
        "type": "string",
        "enum": [
          "repairRequest",
          "payment",
          "certificate",
          "conversation",
          "schedule",
          "address",
          "userDevice",
          "system"
        ]
      },
      "NotificationUrgency": {
        "type": "string",
        "enum": [
          "low",
          "normal",
          "high",
          "critical"
        ]
      },
      "NotificationRecordDto": {
        "type": "object",
        "properties": {
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationType"
              }
            ]
          },
          "targetType": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationTargetType"
              }
            ]
          },
          "urgency": {
            "default": "normal",
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationUrgency"
              }
            ]
          },
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "targetId": {
            "type": "string"
          },
          "metadata": {
            "type": "string"
          },
          "isRead": {
            "type": "boolean"
          },
          "readAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "urgency",
          "id",
          "userId",
          "title",
          "body",
          "isRead",
          "createdAt"
        ]
      },
      "PaginatedNotificationsResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationRecordDto"
            }
          },
          "overallCount": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "limit": {
            "type": "number"
          }
        },
        "required": [
          "data",
          "overallCount",
          "page",
          "limit"
        ]
      },
...
      "NotificationGroup": {
        "type": "string",
        "enum": [
          "repair",
          "payment",
          "schedule",
          "chat",
          "certificate",
          "validation",
          "system"
        ]
      },
      "GroupPreferenceDto": {
        "type": "object",
        "properties": {
          "group": {
            "allOf": [
              {
                "$ref": "#/components/schemas/NotificationGroup"
              }
            ]
          },
          "in_app": {
            "type": "boolean"
          },
          "push": {
            "type": "boolean"
          },
          "email": {
            "type": "boolean"
          }
        },
        "required": [
          "group",
          "in_app",
          "push",
          "email"
        ]
      },
      "NotificationPreferencesResponseDto": {
        "type": "object",
        "properties": {
          "groups": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GroupPreferenceDto"
            }
          },
          "globalMute": {
            "type": "boolean"
          }
        },
        "required": [
          "groups",
          "globalMute"
        ]
      },
      "UpdateNotificationPreferencesDto": {
        "type": "object",
        "properties": {}
      },
      "RegisterPushSubscriptionDto": {
        "type": "object",
        "properties": {}
      },
      "PushSubscriptionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "endpoint",
          "createdAt"
        ]
      },
      "PushSubscriptionListResponseDto": {
        "type": "object",
        "properties": {
          "subscriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PushSubscriptionResponseDto"
            }
          }
        },
        "required": [
          "subscriptions"
        ]
      },
...
      "AddressRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "district": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "house": {
            "type": "string"
          },
          "building": {
            "type": "string"
          },
          "apartment": {
            "type": "string"
          },
          "entrance": {
            "type": "string"
          },
          "floor": {
            "type": "string"
          },
          "intercom": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "validationStatus": {
            "type": "string"
          },
          "validationError": {
            "type": "string"
          },
          "isPrimary": {
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "city",
          "street",
          "house",
          "isPrimary"
        ]
      },
...
      "CreateAddressDto": {
        "type": "object",
        "properties": {}
      },
      "AddressResponseDto": {
        "type": "object",
        "properties": {
          "address": {
            "$ref": "#/components/schemas/AddressRecordDto"
          }
        },
        "required": [
          "address"
        ]
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {}
      },
      "AddressListResponseDto": {
        "type": "object",
        "properties": {
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddressRecordDto"
            }
          }
        },
        "required": [
          "addresses"
        ]
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    }
  }
}